requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.15
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.7
psycopg2-binary==2.9.9
python-decouple==3.8
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
